+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### 2023-04-10
+- Update renderTitle (user ghost buttons)
+- Remove the sort-button in canvas/bookmarks
+- Changed workflow (only trigger on release push)
+- Add option in menu to activate/deactivate the overlay
+- Make canvas overlay width configurable
+- Redesign task manager UI page
+
 ### 2023-04-09
 - Fixed the collapsed window width bug
 - Message when script downloading is started